Former-commit-id: cb9bc980b23f9f8d9cf5276ad9d775e1008d94ea

+++ horace_core/sqw/@PixelData/PixelData.m
@@ -599,7 +599,7 @@ methods
         if isempty(obj.f_accessor_)
             num_pix = size(obj.data, 2);
         else
-            num_pix = obj.f_accessor_.npixels;
+            num_pix = double(obj.f_accessor_.npixels);
         end
     end
